+++ frontend/src/pages/RadioHistory.tsx
@@ -4,12 +4,12 @@ import styled from "styled-components";
 import { device } from "@/constants/styles";
 
 import HeroArea from "@/components/atoms/HeroArea";
-import RadioSearcher from "@/components/atoms/RadioSearcher";
 import AboutSidebar from "@/components/atoms/Features/AboutSidebar";
 import TweetStream from "@/components/atoms/Features/TweetStream";
 
 import PopularRadioWrapper from "@/components/molecules/PopularRadio/PopularRadioWrapper";
 import RadioHistoryWrapper from "@/components/molecules/RadioHistory/RadioHistoryWrapper";
+import RadioSearcher from "@/components/molecules/RadioSearcher";
 import BlogWrapper from "@/components/molecules/Blogs/BlogWrapper";
 
 export default () => (
